Move managers to io
