+++ cartoframes/data/clients/sql_client.py
@@ -1,4 +1,4 @@
-from ...utils.managers.context_manager import ContextManager
+from ...io.managers.context_manager import ContextManager
 
 
 class SQLClient:
